pgplot: Trap out of bounds error in grpxre.f
Sometimes we seem to get an array bound that is one larger than
the size of the array. Seems to be a rounding error in PGPLOT
but no time to debug properly and not always repeatable. For now
trap this and issue a warning rather than continuing with a SEGV.