more elegant solution to forcing linking library with fortran compiler
On OSX 10.7, the previous approach of linking against a dummy fortran
routine the GNU libtool and the system libtool were arguing over a
--force_load option which the GNU libtool insisted the Apple libtool
should accept. It didn't.  Experiments showed that linking
without the dummy routine worked.

The current approach is to force GNU libtool to link using the F77 tag.