interpolate void* so %VAL works.  don't compile %VAL code on buggy gfortran.
The original code use a Fortran preprocessor to insert sizeof(void*) into
the Fortran subroutines which used %VAL.  The AC_PROG_FPP autoconf macro
isn't a standard one, and may be in the Starlink code.  I don't have
much experience with Fortran preprocessors, so this patch simply
uses autoconf's template mechanism to generate .f files from .f.in files.

It also seems that gfortran prior to 4.3 can't handle pointer sizes which
are larger than the default integer size.  I know of no workaround,
so the routines using %VAL are simply not compiled for that compiler.