use pkg-config to find png libraries if possible
on some systems with multiple versions of png installed, sometimes
the versions of library and header are mixed up, to great comic relief.

the pkg-config metadata files use version specific paths, so if they
are available, using them ameliorates the problem.

make sure to update our own pkg-config metadata files to reflect this.