get automake to link convenience libraries with Fortran, take 3
The automake docs recommend using a dummy compilation unit in the
language to be linked by to fool automake into using that language
to link convenience libraries.  The original approach used that concept,
but bundled actual code, which cause problems with the OS X libtool.
Revert to that approach, but configure things so that no code is
required; this makes everyone happy and doesn't require overriding
the libtool tag.