fix breakage under homebrew; streamline testing
On MacOS
1. homebrew does not explicitly add its library directory to the search
   path for linking..

2. By default the system clang stands in for gcc

3. if the homebrew prefix is /usr/local, clang will pick up X11
   libraries and includes from there without prompting, and configure
   will conclude that they exist.

4. homebrew's gfortran (as linker) does not look at /usr/local/lib, so
   will fail to find the X11 libraries.

To solve this, configure now checks for the presence of the 'brew' executable and
prepends $(brew --prefix)/lib to the start of LDFLAGS.

Adding the extra values to test homebrew to the test matrix exploded
the number of configurations to no great win, as only MacOS needs
those flags, and the other platforms would have had to be explicitly
excluded.  Instead, there's now one job per platform, with the common
code split into an action (build-pgplot).  Each platform has their
own matrix of options